renamed SignUp components to Login and changed the display 'Sign up' in navbar and footer to 'Login'
