+++ src/index.css
@@ -423,7 +423,7 @@ body::-webkit-scrollbar-thumb {
   margin-top: 8%;
 }
 
-/*-- SignUp Page --*/
+/*-- Login Page --*/
 
 .signin-margin {
   margin-top: 100px;
